+++ src/main/kotlin/restApi/VideoMessage.kt
@@ -11,7 +11,7 @@ fun routeVideoMessage(routing: Routing): Unit = with(routing) {
         route("video-message") {
             getMediaMessage(this) { messageId -> VideoMessages.read(messageId).bytes }
             postMediaMessage(this, { readMultipartMp4() }) { userId, chatId, message, contextMessageId ->
-                Messages.createVideoMessage(userId, chatId, message, contextMessageId)
+                Messages.createVideoMessage(userId, chatId, message, contextMessageId, isForwarded = false)
             }
         }
     }