+++ src/test/kotlin/db/tables/MessagesUtil.kt
@@ -3,17 +3,23 @@ package com.neelkamath.omniChat.db.tables
 import com.neelkamath.omniChat.graphql.routing.MessageText
 import com.neelkamath.omniChat.graphql.routing.PollInput
 
-fun Messages.create(userId: Int, chatId: Int, text: MessageText = MessageText("t")): Unit =
-    createTextMessage(userId, chatId, text, contextMessageId = null)
+fun Messages.create(
+    userId: Int,
+    chatId: Int,
+    text: MessageText = MessageText("t"),
+    contextMessageId: Int? = null,
+    isForwarded: Boolean = false
+): Unit = createTextMessage(userId, chatId, text, contextMessageId, isForwarded)
 
 /** Sends the [message] in the [chatId] from the [userId], and returns the message's ID. */
 fun Messages.message(
     userId: Int,
     chatId: Int,
     message: MessageText = MessageText("t"),
-    contextMessageId: Int? = null
+    contextMessageId: Int? = null,
+    isForwarded: Boolean = false
 ): Int {
-    createTextMessage(userId, chatId, message, contextMessageId)
+    createTextMessage(userId, chatId, message, contextMessageId, isForwarded)
     return readIdList(chatId).last()
 }
 
@@ -21,9 +27,10 @@ fun Messages.message(
     userId: Int,
     chatId: Int,
     message: Mp3,
-    contextMessageId: Int? = null
+    contextMessageId: Int? = null,
+    isForwarded: Boolean = false
 ): Int {
-    createAudioMessage(userId, chatId, message, contextMessageId)
+    createAudioMessage(userId, chatId, message, contextMessageId, isForwarded)
     return readIdList(chatId).last()
 }
 
@@ -31,9 +38,10 @@ fun Messages.message(
     userId: Int,
     chatId: Int,
     message: CaptionedPic,
-    contextMessageId: Int? = null
+    contextMessageId: Int? = null,
+    isForwarded: Boolean = false
 ): Int {
-    createPicMessage(userId, chatId, message, contextMessageId)
+    createPicMessage(userId, chatId, message, contextMessageId, isForwarded)
     return readIdList(chatId).last()
 }
 
@@ -41,8 +49,9 @@ fun Messages.message(
     userId: Int,
     chatId: Int,
     message: PollInput,
-    contextMessageId: Int? = null
+    contextMessageId: Int? = null,
+    isForwarded: Boolean = false
 ): Int {
-    createPollMessage(userId, chatId, message, contextMessageId)
+    createPollMessage(userId, chatId, message, contextMessageId, isForwarded)
     return readIdList(chatId).last()
 }